Fix eval_commons: parse per-layer MLP mult as float, not int
Tapered models (e.g. 322 pattern) use fractional MLP multipliers
like 0.5. Matches the fix already in train_gpt_mlx.py.

+++ scripts/eval_commons.py
@@ -41,7 +41,7 @@ def build_model(hparams: Hyperparameters) -> GPT:
     """Build a GPT model from Hyperparameters. Replaces ~15 inline copies."""
     per_layer = None
     if hparams.mlp_mult_per_layer:
-        per_layer = [int(x) for x in hparams.mlp_mult_per_layer.split(",")]
+        per_layer = [float(x) for x in hparams.mlp_mult_per_layer.split(",")]
     return GPT(
         vocab_size=hparams.vocab_size, num_layers=hparams.num_layers,
         dim=hparams.model_dim, num_heads=hparams.num_heads,
